+++ SLK_img2pixel/src/process.h
@@ -36,7 +36,7 @@ extern int dither_amount;
 extern int alpha_threshold;
 extern int gauss;
 
-void process_image(const SLK_RGB_sprite *in, SLK_RGB_sprite *out, SLK_Palette *palette, int sample_mode, int process_mode);
+void process_image(const SLK_RGB_sprite *in, SLK_RGB_sprite *out, SLK_Palette *palette, int sample_mode, int process_mode, int distance_mode);
 void sharpen_image(SLK_RGB_sprite *in, SLK_RGB_sprite *out);
 void lowpass_image(SLK_RGB_sprite *in, SLK_RGB_sprite *out);