+++ SLK_img2pixel/src/utility.h
@@ -21,8 +21,8 @@ void palette_write(SLK_Palette *pal);
 
 SLK_RGB_sprite *image_load(const char *path);
 void dir_input_select();
-void dir_output_select(int dither_mode, int sample_mode, int scale_mode, int width, int height, SLK_Palette *pal);
+void dir_output_select(int dither_mode, int sample_mode, int distance_mode, int scale_mode, int width, int height, SLK_Palette *pal);
 void gif_input_select();
-void gif_output_select(int dither_mode, int sample_mode, int scale_mode, int width, int height, SLK_Palette *pal);
+void gif_output_select(int dither_mode, int sample_mode, int distance_mode, int scale_mode, int width, int height, SLK_Palette *pal);
 
 #endif